chore: remove dndTreeSelect
