+++ src/components/Table/hoc/withTableSettings/TableColumnSetup/TableColumnSetup.tsx
@@ -1,17 +1,27 @@
 import React from 'react';
 
-import {Gear, Lock} from '@gravity-ui/icons';
+import {Gear, Grip, Lock} from '@gravity-ui/icons';
+import {DragDropContext, Draggable, Droppable} from 'react-beautiful-dnd';
+import type {OnDragEndResponder} from 'react-beautiful-dnd';
 
+import type {
+    RenderContainerType,
+    RenderItem,
+    TreeSelectProps,
+} from 'src/components/TreeSelect/types';
+
+import {useUniqId} from '../../../../../hooks';
 import type {PopperPlacement} from '../../../../../hooks/private';
 import {createOnKeyDownHandler} from '../../../../../hooks/useActionHandlers/useActionHandlers';
 import {Button} from '../../../../Button';
 import {Icon} from '../../../../Icon';
-import type {TreeSelectProps} from '../../../../TreeSelect';
+import {TreeSelect} from '../../../../TreeSelect/TreeSelect';
+import {TreeSelectItem} from '../../../../TreeSelect/TreeSelectItem';
+import type {TreeSelectItemProps} from '../../../../TreeSelect/TreeSelectItem';
+import {ListContainerView} from '../../../../useList';
 import {block} from '../../../../utils/cn';
 import type {TableColumnSetupItem, TableSetting} from '../withTableSettings';
 
-import {DndTreeSelect} from './DndTreeSelect';
-import type {DndTreeSelectItem, DndTreeSelectProps, RenderDndContainer} from './DndTreeSelect';
 import i18n from './i18n';
 
 import './TableColumnSetup.scss';
@@ -21,12 +31,24 @@ const tableColumnSetupCn = b(null);
 const applyButtonCn = b('apply');
 const requiredDndItemCn = b('required-item');
 
+const reorderArray = <T extends unknown>(list: T[], startIndex: number, endIndex: number): T[] => {
+    const result = [...list];
+    const [removed] = result.splice(startIndex, 1);
+    result.splice(endIndex, 0, removed);
+
+    return result;
+};
+
 interface SwitcherProps {
     onKeyDown: React.KeyboardEventHandler<HTMLElement>;
     onClick: React.MouseEventHandler<HTMLElement>;
 }
 
-type Item = TableColumnSetupItem & DndTreeSelectItem;
+type Item = TableColumnSetupItem &
+    TreeSelectItemProps & {
+        id: string;
+        isDragDisabled?: boolean;
+    };
 
 export interface TableColumnSetupProps {
     renderSwitcher?: (props: SwitcherProps) => React.JSX.Element;
@@ -35,7 +57,7 @@ export interface TableColumnSetupProps {
     sortable?: boolean;
 
     onUpdate: (newSettings: TableSetting[]) => void;
-    popupWidth?: DndTreeSelectProps<any>['popupWidth'];
+    popupWidth?: TreeSelectProps<any>['popupWidth'];
     popupPlacement?: PopperPlacement;
 }
 
@@ -48,6 +70,8 @@ export const TableColumnSetup = (props: TableColumnSetupProps) => {
         onUpdate: propsOnUpdate,
         sortable,
     } = props;
+
+    const uniqId = useUniqId();
 
     const [open, setOpen] = React.useState(false);
 
@@ -83,17 +107,6 @@ export const TableColumnSetup = (props: TableColumnSetupProps) => {
         setOpen(false);
     };
 
-    const renderContainer: RenderDndContainer = ({renderList}) => {
-        return (
-            <React.Fragment>
-                {renderList()}
-                <Button view="action" className={applyButtonCn} onClick={onApply}>
-                    {i18n('button_apply')}
-                </Button>
-            </React.Fragment>
-        );
-    };
-
     const renderControl: TreeSelectProps<unknown>['renderControl'] = ({toggleOpen}) => {
         const onKeyDown = createOnKeyDownHandler(toggleOpen);
 
@@ -116,13 +129,98 @@ export const TableColumnSetup = (props: TableColumnSetupProps) => {
         }
     };
 
-    const onUpdate = React.useCallback((selectedItemsIds: string[]) => {
+    const onUpdate = (selectedItemsIds: string[]) => {
         setItems((prevItems) => {
             return prevItems.map((item) => ({
                 ...item,
                 isSelected: selectedItemsIds.includes(item.id),
             }));
         });
+    };
+
+    const renderContainer = React.useCallback<RenderContainerType<Item>>(
+        ({renderItem, visibleFlattenIds, items: _items, containerRef, id}) => {
+            const handleDrugEnd: OnDragEndResponder = ({destination, source}) => {
+                if (destination?.index !== undefined && destination?.index !== source.index) {
+                    setItems((prevItems) => {
+                        return reorderArray(prevItems, source.index, destination.index);
+                    });
+                }
+            };
+
+            const visibleFlattenItemList = visibleFlattenIds.map((visibleFlattenId, idx) =>
+                renderItem(visibleFlattenId, idx),
+            );
+
+            return (
+                <React.Fragment>
+                    <ListContainerView ref={containerRef} id={id}>
+                        <DragDropContext onDragEnd={handleDrugEnd}>
+                            <Droppable droppableId={uniqId}>
+                                {(droppableProvided) => {
+                                    return (
+                                        <div
+                                            {...droppableProvided.droppableProps}
+                                            ref={droppableProvided.innerRef}
+                                        >
+                                            {visibleFlattenItemList}
+                                            {droppableProvided.placeholder}
+                                        </div>
+                                    );
+                                }}
+                            </Droppable>
+                        </DragDropContext>
+                    </ListContainerView>
+                    <Button view="action" className={applyButtonCn} onClick={onApply}>
+                        {i18n('button_apply')}
+                    </Button>
+                </React.Fragment>
+            );
+        },
+        [],
+    );
+
+    const renderItem = React.useCallback<RenderItem<Item>>((item, state, _, idx) => {
+        const endSlot =
+            item.endSlot ?? (item.isDragDisabled ? undefined : <Icon data={Grip} size={16} />);
+
+        const commonProps = {
+            ...state,
+            ...item,
+            endSlot,
+        };
+
+        return (
+            <Draggable
+                draggableId={state.id}
+                index={Number(idx)}
+                key={`item-key-${state.id}`}
+                isDragDisabled={item.isDragDisabled}
+            >
+                {(provided, snapshot) => {
+                    const style: React.CSSProperties = {
+                        ...provided.draggableProps.style,
+                    };
+
+                    // not expected offset appears, one way to fix - remove this offsets explicitly
+                    if (snapshot.isDragging) {
+                        style.left = undefined;
+                        style.top = undefined;
+                    }
+
+                    return (
+                        <TreeSelectItem
+                            ref={provided.innerRef}
+                            {...commonProps}
+                            {...provided.draggableProps}
+                            {...provided.dragHandleProps}
+                            style={style}
+                            active={snapshot.isDragging}
+                        />
+                    );
+                }}
+            </Draggable>
+        );
     }, []);
 
     const value = React.useMemo(() => {
@@ -138,7 +236,7 @@ export const TableColumnSetup = (props: TableColumnSetupProps) => {
     }, [items]);
 
     return (
-        <DndTreeSelect
+        <TreeSelect
             className={tableColumnSetupCn}
             multiple
             size="l"
@@ -150,8 +248,8 @@ export const TableColumnSetup = (props: TableColumnSetupProps) => {
             placement={popupPlacement}
             renderContainer={renderContainer}
             renderControl={renderControl}
+            renderItem={renderItem}
             items={items}
-            setItems={setItems}
         />
     );
 };
